left note about YPred column
Matlab auto assigns output in the network classification layer since our code did not specify an order. I have since discovered that this can change based on the version of matlab and/or linux vs. windows implementation. I recommend specifying "Classes" in the output layer of the network to avoid confusion. Unfortunately, we learned this after the fact.